Protected routes (#32)
* config : set accessible routes with certain authentication roles

* fix : condition the recovery of the exception message for ConstraintViolationException

* feat : check that entities belong to the logged in user

+++ src/main/java/com/jeyofdev/yellow_berry/domain/profile/ProfileService.java
@@ -5,12 +5,17 @@ import com.jeyofdev.yellow_berry.auth_user.AuthUserService;
 import com.jeyofdev.yellow_berry.core.classes.AbstractDomainService;
 import com.jeyofdev.yellow_berry.core.constant.ConfirmMessage;
 import com.jeyofdev.yellow_berry.core.constant.ErrorMessage;
+import com.jeyofdev.yellow_berry.domain.product.Product;
+import com.jeyofdev.yellow_berry.domain.wishlist.WishList;
 import com.jeyofdev.yellow_berry.exception.AlreadyTakenException;
 import com.jeyofdev.yellow_berry.exception.AlreadyAssociatedException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.text.MessageFormat;
+import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -26,41 +31,57 @@ public class ProfileService extends AbstractDomainService<Profile, ProfileReposi
     }
 
     public Profile save(UUID userId, Profile profile) {
+        String username  = SecurityContextHolder.getContext().getAuthentication().getName();
         AuthUser user = authUserService.findUserById(userId);
 
-        if (user.getProfile() != null) {
-            throw new AlreadyAssociatedException(MessageFormat.format(ErrorMessage.ALREADY_ASSOCIATED, "user", "profile"));
+        if (username.equals(user.getUsername())) {
+            if (user.getProfile() != null) {
+                throw new AlreadyAssociatedException(MessageFormat.format(ErrorMessage.ALREADY_ASSOCIATED, "user", "profile"));
+            }
+
+            if (profileRepository.existsByPhone(profile.getPhone())) {
+                throw new AlreadyTakenException(MessageFormat.format(ErrorMessage.ALREADY_TAKEN, entityName, "phone", profile.getPhone()));
+            }
+
+            profile.setUser(user);
+
+            return profileRepository.save(profile);
+        } else {
+            throw new AccessDeniedException(ErrorMessage.LIMIT_ACCESS);
         }
-
-        if (profileRepository.existsByPhone(profile.getPhone())) {
-            throw new AlreadyTakenException(MessageFormat.format(ErrorMessage.ALREADY_TAKEN, entityName, "phone", profile.getPhone()));
-        }
-
-
-        profile.setUser(user);
-
-        return profileRepository.save(profile);
     }
 
     public Profile updateById(UUID profileId, Profile updatedProfile) {
+        String username  = SecurityContextHolder.getContext().getAuthentication().getName();
         Profile existingProfile = findById(profileId);
 
-        existingProfile.setFirstname(updatedProfile.getFirstname() != null ? updatedProfile.getFirstname() : existingProfile.getFirstname());
-        existingProfile.setLastname(updatedProfile.getLastname() != null ? updatedProfile.getLastname() : existingProfile.getLastname());
-        existingProfile.setPhone(updatedProfile.getPhone() != null ? updatedProfile.getPhone() : existingProfile.getPhone());
-        existingProfile.setAddress(updatedProfile.getAddress() != null ? updatedProfile.getAddress() : existingProfile.getAddress());
-        existingProfile.setRegion(updatedProfile.getRegion() != null ? updatedProfile.getRegion() : existingProfile.getRegion());
-        existingProfile.setDepartment(updatedProfile.getDepartment() != null ? updatedProfile.getDepartment() : existingProfile.getDepartment());
-        existingProfile.setZipCode(updatedProfile.getZipCode() != null ? updatedProfile.getZipCode() : existingProfile.getZipCode());
-        existingProfile.setCity(updatedProfile.getCity() != null ? updatedProfile.getCity() : existingProfile.getCity());
+        if (username.equals(existingProfile.getUser().getUsername())) {
+            existingProfile.setFirstname(updatedProfile.getFirstname() != null ? updatedProfile.getFirstname() : existingProfile.getFirstname());
+            existingProfile.setLastname(updatedProfile.getLastname() != null ? updatedProfile.getLastname() : existingProfile.getLastname());
+            existingProfile.setPhone(updatedProfile.getPhone() != null ? updatedProfile.getPhone() : existingProfile.getPhone());
+            existingProfile.setAddress(updatedProfile.getAddress() != null ? updatedProfile.getAddress() : existingProfile.getAddress());
+            existingProfile.setRegion(updatedProfile.getRegion() != null ? updatedProfile.getRegion() : existingProfile.getRegion());
+            existingProfile.setDepartment(updatedProfile.getDepartment() != null ? updatedProfile.getDepartment() : existingProfile.getDepartment());
+            existingProfile.setZipCode(updatedProfile.getZipCode() != null ? updatedProfile.getZipCode() : existingProfile.getZipCode());
+            existingProfile.setCity(updatedProfile.getCity() != null ? updatedProfile.getCity() : existingProfile.getCity());
 
-        return profileRepository.save(existingProfile);
+            return profileRepository.save(existingProfile);
+        } else {
+            throw new AccessDeniedException(ErrorMessage.LIMIT_ACCESS);
+        }
     }
 
     public String deleteById(UUID profileId) {
-        findById(profileId);
-        profileRepository.deleteById(profileId);
+        String username  = SecurityContextHolder.getContext().getAuthentication().getName();
+        String roles  = SecurityContextHolder.getContext().getAuthentication().getAuthorities().toString();
+        Profile profile = findById(profileId);
 
-        return ConfirmMessage.PROFILE_DELETE;
+        if (username.equals(profile.getUser().getUsername()) || roles.equals("[ROLE_ADMIN]")) {
+            profileRepository.deleteById(profileId);
+
+            return ConfirmMessage.PROFILE_DELETE;
+        } else {
+            throw new AccessDeniedException(ErrorMessage.LIMIT_ACCESS);
+        }
     }
 }
